Fleshed out more room functions
Completely finished the Rooms_list dict -- this allowed me to flesh out more of the room functions, which are called any time a valid direction choice is made. This is a much better format than calling what was basically a copy of each room for every time an invalid direction was chosen.